added getFields to retrieve fields
Added the `getFields` method on DynamicEntry to retrieve fields. This gives us the opportunity to check whether properties exist or not. The problem ist, that currently I don't see any possibility to keep track on non existent properties (no magic method). 

Would be great to merge that.
Thanks.

Otherwise we could add a method to check for existing properties (eg: `propertyExist($name)`)