Adds the possibility to query for a single locale.
Also make it possible to specify a default locale when creating a client.

This is a major breaking change but adresses an often-requested feature. The new default
behavior is is to request content the space's default locale. To retain the old behavior
set the default locale to `'*''` when created the client. This could look like this:
`new Client($token, $spaceID, false, null, null, '*')`