Initial php8 support
This commit adds the initial support and changes required for php8.
There are several things of note here. The smallest change is the update
to the README and the CHANGELOG, as required. The version will be bumped
to 6.1.0, as this adds a new feature, and the new version was added to
the composer.json.

The next larger change is the update to the CI process. It does not
differ much from the changes to the core lib: A CI job for 8.0 was added
and the docs build was fixed to avoid the SAMI deprecation. One library
required for testing, namely php-vcr, does not support php8 yet. This is
a bit problematic, as the unit tests (especially the failure ones) rely
on it. This is the only library with this problem and is therefore
force-installed via --ignore-platform-reqs; since it is only used for
unit testing, this should be good for now.

Lastly, an unit test which checked the PHP ini error reporting setting
was disabled, as it caused test failures on newer images for no reason.